requirements.txt
certifi==2019.6.16
Django==2.2.5
django-common-helpers==0.9.2
django-cron==0.5.1
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0
